Implemented basic material reflection

+++ src/patterns/checker.cpp
@@ -2,7 +2,7 @@
 
 const Color Checker::colorAt(const Tuple& p) const
 {
-    int dimensionSum = floor(p._x) + floor(p._y) + floor(p._z);
+    int dimensionSum = (int)floor(p._x) + (int)floor(p._y) + (int)floor(p._z);
     if (dimensionSum % 2 == 0)
         return _fst->colorAt(p);
     return _scd->colorAt(p);
